+++ packages/excalidraw/components/Stats/MultiAngle.tsx
@@ -9,8 +9,8 @@ import DragInput from "./DragInput";
 import type { DragInputCallbackType } from "./DragInput";
 import { getStepSizedValue, isPropertyEditable } from "./utils";
 import type { AppState } from "../../types";
-import type { Degrees } from "@excalidraw/math";
-import { degreesToRadians, radiansToDegrees } from "@excalidraw/math";
+import type { Degrees } from "../../../math";
+import { degreesToRadians, radiansToDegrees } from "../../../math";
 
 interface MultiAngleProps {
   elements: readonly ExcalidrawElement[];
