+++ packages/excalidraw/components/Stats/utils.ts
@@ -1,5 +1,5 @@
-import type { Radians } from "@excalidraw/math";
-import { point, pointRotateRads } from "@excalidraw/math";
+import type { Radians } from "../../../math";
+import { point, pointRotateRads } from "../../../math";
 import {
   bindOrUnbindLinearElements,
   updateBoundElements,
