+++ packages/excalidraw/element/bounds.ts
@@ -30,7 +30,7 @@ import type {
   LineSegment,
   LocalPoint,
   Radians,
-} from "@excalidraw/math";
+} from "../../math";
 import {
   degreesToRadians,
   lineSegment,
@@ -38,7 +38,7 @@ import {
   pointDistance,
   pointFromArray,
   pointRotateRads,
-} from "@excalidraw/math";
+} from "../../math";
 import type { Mutable } from "../utility-types";
 
 export type RectangleBox = {
